(Windows) Move .mo files to "share" subdir under python root dir
Move the .mo files to the $PYTHONDIR/share subdir when installing openmolar on Windows, so that the translations work without much effort (also, on windows, the user can set the LANG environment variable to change the language easily). I'm not really sure if this is correct, or a good idea, though...
This comes from the realization that gettext.bindtextdomain('openmolar') returns this path under Windows, while it returns '/usr/share/locale' in Linux.